requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
openpyxl==3.1.2
python-Levenshtein==0.23.0
pydantic==2.5.0
aiofiles==23.2.1
anthropic==0.39.0
python-dotenv==1.0.0
